requirements.txt
requests==2.32.3
anthropic==0.49.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
